linting and cleanup
